refactor: convert Item class to interface
